Extract topic logic into helper
