+++ src/main/scala/changestream/actors/SnsActor.scala
@@ -4,6 +4,7 @@ import scala.language.postfixOps
 import scala.util.{Failure, Success}
 import akka.actor.Actor
 import changestream.events.{MutationEvent, MutationWithInfo}
+import changestream.helpers.Topic
 import com.amazonaws.services.sns.AmazonSNSAsyncClient
 import com.amazonaws.services.sns.model.CreateTopicResult
 import com.github.dwhjames.awswrap.sns.AmazonSNSScalaClient
@@ -12,17 +13,6 @@ import org.slf4j.LoggerFactory
 
 import scala.collection.mutable
 import scala.concurrent.Future
-
-object SnsActor {
-  def getTopic(mutation: MutationEvent, topic: String, topicHasVariable: Boolean = true): String = {
-    val database = mutation.database.replaceAll("[^a-zA-Z0-9\\-_]", "-")
-    val tableName = mutation.tableName.replaceAll("[^a-zA-Z0-9\\-_]", "-")
-    topicHasVariable match {
-      case true => topic.replace("{database}", database).replace("{tableName}", tableName)
-      case false => topic
-    }
-  }
-}
 
 class SnsActor(config: Config = ConfigFactory.load().getConfig("changestream")) extends Actor {
   protected val log = LoggerFactory.getLogger(getClass)
@@ -47,7 +37,7 @@ class SnsActor(config: Config = ConfigFactory.load().getConfig("changestream")) 
       log.debug(s"Received message: ${message}")
 
       val origSender = sender()
-      val topic = SnsActor.getTopic(mutation, snsTopic, snsTopicHasVariable)
+      val topic = Topic.getTopic(mutation, snsTopic, snsTopicHasVariable)
       val topicArn = topicArns.getOrElse(topic, client.createTopic(topic))
       topicArns.update(topic, topicArn)
